Refactor: Improve angle precision and clean up code.
This commit increases the precision of angular calculations by using `double` instead of `float` and refactors related utility methods and UI code.

- Changed `getAzimuthDeg()` in `SunEvent` and `SunAzimuthInfo` to return `double` for higher precision.
- Replaced `DirectionUtil.zeroCenterDeg(double)` with `DirectionUtil.arc(double, double)`, which calculates the shortest signed angular distance using `IEEErem` for better accuracy and clarity.
- Added `DirectionUtilTest.kt` to provide comprehensive test coverage for the new `arc` function.
- Updated `CompassActivity` to use the new `DirectionUtil.arc` method for determining noon/nadir overlap and simplified the compass rotation animation logic.
- Removed programmatic setting of `iconTintList` in `BaseActivity` and moved the tinting logic to `actions_menu.xml` using `app:iconTint`.
- Renamed `styles.xml` to `themes.xml` to follow modern Android conventions.
- Added `truth` and other testing dependencies to `build.gradle.kts`.

+++ helios/src/main/java/org/onereed/helios/common/LayoutParamsUtil.java
@@ -13,8 +13,8 @@ public class LayoutParamsUtil {
   }
 
   /** Sets the angle of a view that is using circle-constrained layout. */
-  public static void changeConstraintLayoutCircleAngle(View view, float angle) {
-    changeConstraintLayoutParams(view, layoutParams -> layoutParams.circleAngle = angle);
+  public static void changeConstraintLayoutCircleAngle(View view, double angle) {
+    changeConstraintLayoutParams(view, layoutParams -> layoutParams.circleAngle = (float) angle);
   }
 
   /**
